+++ src/lib/orionld/rest/orionldServiceInit.cpp
@@ -1,0 +1,139 @@
+/*
+*
+* Copyright 2018 Telefonica Investigacion y Desarrollo, S.A.U
+*
+* This file is part of Orion Context Broker.
+*
+* Orion Context Broker is free software: you can redistribute it and/or
+* modify it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* Orion Context Broker is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+* General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with Orion Context Broker. If not, see http://www.gnu.org/licenses/.
+*
+* For those usages not covered by this license please contact with
+* iot_support at tid dot es
+*
+* Author: Ken Zangelin
+*/
+#include <microhttpd.h>
+
+#include "orionld/rest/OrionLdRestService.h"          // OrionLdRestService
+#include "orionld/rest/temporaryErrorPayloads.h"      // Temporary Error Payloads
+#include "orionld/rest/orionldMhdConnection.h"        // Own Interface
+
+
+
+// -----------------------------------------------------------------------------
+//
+// orionldRestServiceV -
+//
+OrionLdRestServiceVector orionldRestServiceV[9];
+
+
+
+// -----------------------------------------------------------------------------
+//
+// restServicePrepare -
+//
+static void restServicePrepare(OrionLdRestService* serviceP, OrionLdRestServiceSimplified* simpleServiceP)
+{
+  // 1. Simply copy/reference the two fields of OrionLdRestServiceSimplified
+  serviceP->url             = (char*) simpleServiceP->url;
+  serviceP->serviceRoutine  = simpleServiceP->serviceRoutine;
+
+  // 2. Loop over the URL Path, count wildcards, charsBeforeFirstWildcard, etc
+  int    ix            = 11;    // strlen("/ngsi-ld/v1/") - 1
+  char*  wildCardStart = NULL;
+  char*  wildCardEnd   = NULL;
+
+  while (serviceP->url[++ix] != 0)
+  {
+    char c = serviceP->url[ix];
+
+    if (c == '*')
+    {
+      if (serviceP->wildcards == 0)
+        wildCardStart = &serviceP->url[ix + 1];
+      else if (serviceP->wildcards == 1)
+        wildCardEnd = &serviceP->url[ix];
+        
+      LM_TMP(("Found a wildcard in index %d of '%s'", ix, serviceP->url));
+      serviceP->wildcards += 1;
+      continue;
+    }
+
+    if (serviceP->wildcards == 0)
+    {
+      ++serviceP->charsBeforeFirstWildcard;
+      serviceP->charsBeforeFirstWildcardSum += c;
+    }
+    else if (serviceP->wildcards == 1)
+    {
+      ++serviceP->charsBeforeSecondWildcard;
+      serviceP->charsBeforeSecondWildcardSum += c;
+    }
+  }
+
+  if (serviceP->wildcards != 0)
+  {
+    if (wildCardEnd == NULL)  // If only one '*', make wildCardEnd point to end of URL
+      wildCardEnd = &serviceP->url[ix];
+
+    serviceP->matchForSecondWildcardLen = wildCardEnd - wildCardStart;
+
+    if (serviceP->matchForSecondWildcardLen < 0)
+      LM_X(1, ("Negative length of matchForSecondWildcard - not possible. SW bug"));
+    if (serviceP->matchForSecondWildcardLen > (int) sizeof(serviceP->matchForSecondWildcard))
+      LM_X(1, ("Too big matchForSecondWildcard - not possible. SW bug"));
+
+    if (serviceP->matchForSecondWildcardLen != 0)
+    {
+      strncpy(serviceP->matchForSecondWildcard, wildCardStart, wildCardEnd - wildCardStart);
+      LM_TMP(("matchForSecondWildcard: %s", serviceP->matchForSecondWildcard));
+    }
+  }
+}
+
+
+
+// -----------------------------------------------------------------------------
+//
+// orionLdServiceInit -
+//
+// This function converts the OrionLdRestServiceSimplified vectors to OrionLdRestService vectors
+//
+void orionLdServiceInit(OrionLdRestServiceSimplifiedVector* restServiceVV, int vecItems)
+{
+  int svIx;  // Service Vector Index
+
+  bzero(orionldRestServiceV, sizeof(orionldRestServiceV));
+
+  for (svIx = 0; svIx < vecItems; svIx++)
+  {
+    // svIx is really the Verb (GET=0, POST=2, up to NOVERB=9. See src/lib/rest/Verb.h)
+
+    LM_TMP(("svIx: %d", svIx));
+    if (restServiceVV[svIx].serviceV == NULL)
+      continue;
+
+    int services = restServiceVV[svIx].services;
+    
+    orionldRestServiceV[svIx].serviceV  = (OrionLdRestService*) calloc(sizeof(OrionLdRestService), services);
+    orionldRestServiceV[svIx].services  = services;
+
+    int sIx;  // Service Index inside Rest Service vector
+
+    for (sIx = 0; sIx < services; sIx++)
+    {
+      LM_TMP(("sIx: %d", sIx));
+      restServicePrepare(&orionldRestServiceV[svIx].serviceV[sIx], &restServiceVV[svIx].serviceV[sIx]);
+    }
+  }
+}
