+++ src/lib/orionld/rest/temporaryErrorPayloads.cpp
@@ -1,0 +1,9 @@
+/* ****************************************************************************
+*
+* Temporary predefined payloads
+*/
+const char* invalidVerbPayload     = "{\"error\":\"Invalid Verb\"}";
+const char* payloadTooLargePayload = "{\"error\":\"Payload too large\"}";
+const char* genericErrorPayload    = "{\"error\":\"Generic Error\"}";
+const char* doubleSlashPayload     = "{\"error\":\"Double Slash in URL PATH\"}";
+const char* notFoundPayload        = "{\"error\":\"Service not found\"}";
