+++ src/lib/orionld/rest/temporaryErrorPayloads.h
@@ -1,0 +1,16 @@
+#ifndef SRC_LIB_ORIONLD_REST_TEMPORARYERRORPAYLOADS_H_
+#define SRC_LIB_ORIONLD_REST_TEMPORARYERRORPAYLOADS_H_
+
+
+
+/* ****************************************************************************
+*
+* Temporary predefined payloads
+*/
+extern const char* invalidVerbPayload;
+extern const char* payloadTooLargePayload;
+extern const char* genericErrorPayload;
+extern const char* doubleSlashPayload;
+extern const char* notFoundPayload;
+
+#endif  // SRC_LIB_ORIONLD_REST_TEMPORARYERRORPAYLOADS_H_
